feat(color-picker): Export new default color set
Revert previous changes to export color set from different file.

Update default color set's neutrals  according to design.
