+++ modules/_labs/color-picker/react/index.ts
@@ -1,2 +1,1 @@
 export {default, default as ColorPicker, ColorPickerProps} from './lib/ColorPicker';
-export {defaultColorSet, fullSaturationColorSet} from './lib/parts/ColorSets';
