+++ modules/_labs/color-picker/react/lib/ColorPicker.tsx
@@ -1,4 +1,4 @@
-import {spacing} from '@workday/canvas-kit-react-core';
+import {colors, spacing} from '@workday/canvas-kit-react-core';
 import {checkIcon} from '@workday/canvas-system-icons-web';
 import {ColorInput} from '@workday/canvas-kit-react-color-picker';
 import {IconButton} from '@workday/canvas-kit-react-button';
@@ -10,7 +10,6 @@ import styled from '@emotion/styled';
 
 import {ResetButton} from './parts/ColorReset';
 import {SwatchBook} from './parts/SwatchBook';
-import {defaultColorSet} from './parts/ColorSets';
 
 export interface ColorPickerProps extends React.HTMLAttributes<HTMLDivElement> {
   /**
@@ -65,6 +64,71 @@ export interface ColorPickerProps extends React.HTMLAttributes<HTMLDivElement> {
    */
   transformOrigin?: TransformOrigin;
 }
+
+export const defaultColorSet = [
+  colors.blueberry600,
+  colors.grapeSoda600,
+  colors.pomegranate600,
+  colors.cinnamon600,
+  colors.cantaloupe600,
+  colors.sourLemon600,
+  colors.greenApple600,
+  colors.jewel600,
+
+  colors.blueberry500,
+  colors.grapeSoda500,
+  colors.pomegranate500,
+  colors.cinnamon500,
+  colors.cantaloupe500,
+  colors.sourLemon500,
+  colors.greenApple500,
+  colors.jewel500,
+
+  colors.blueberry400,
+  colors.grapeSoda400,
+  colors.pomegranate400,
+  colors.cinnamon400,
+  colors.cantaloupe400,
+  colors.sourLemon400,
+  colors.greenApple400,
+  colors.jewel400,
+
+  colors.blueberry300,
+  colors.grapeSoda300,
+  colors.pomegranate300,
+  colors.cinnamon300,
+  colors.cantaloupe300,
+  colors.sourLemon300,
+  colors.greenApple300,
+  colors.jewel300,
+
+  colors.blueberry200,
+  colors.grapeSoda200,
+  colors.pomegranate200,
+  colors.cinnamon200,
+  colors.cantaloupe200,
+  colors.sourLemon200,
+  colors.greenApple200,
+  colors.jewel200,
+
+  colors.blueberry100,
+  colors.grapeSoda100,
+  colors.pomegranate100,
+  colors.cinnamon100,
+  colors.cantaloupe100,
+  colors.sourLemon100,
+  colors.greenApple100,
+  colors.jewel100,
+
+  colors.blackPepper600,
+  colors.blackPepper400,
+  colors.blackPepper300,
+  colors.blackPepper100,
+  colors.frenchVanilla500,
+  colors.frenchVanilla400,
+  colors.frenchVanilla200,
+  colors.frenchVanilla100,
+];
 
 const ColorInputWrapper = styled('form')({
   width: '100%',
