--- modules/_labs/color-picker/react/lib/parts/ColorSets.ts
@@ -1,95 +1,0 @@
-import {colors} from '@workday/canvas-kit-react-core';
-
-export const defaultColorSet = [
-  colors.cinnamon400,
-  colors.peach400,
-  colors.chiliMango400,
-  colors.cantaloupe400,
-  colors.sourLemon400,
-  colors.juicyPear400,
-  colors.kiwi400,
-  colors.greenApple400,
-
-  colors.watermelon400,
-  colors.jewel400,
-  colors.toothpaste400,
-  colors.blueberry400,
-  colors.plum400,
-  colors.berrySmoothie400,
-  colors.blackberry400,
-  colors.islandPunch400,
-
-  colors.grapeSoda400,
-  colors.pomegranate400,
-  colors.fruitPunch400,
-  colors.rootBeer400,
-  colors.toastedMarshmallow400,
-  colors.licorice400,
-  colors.cappuccino400,
-  colors.blackPepper400,
-];
-
-export const fullSaturationColorSet = [
-  colors.blueberry600,
-  colors.grapeSoda600,
-  colors.pomegranate600,
-  colors.cinnamon600,
-  colors.cantaloupe600,
-  colors.sourLemon600,
-  colors.greenApple600,
-  colors.jewel600,
-
-  colors.blueberry500,
-  colors.grapeSoda500,
-  colors.pomegranate500,
-  colors.cinnamon500,
-  colors.cantaloupe500,
-  colors.sourLemon500,
-  colors.greenApple500,
-  colors.jewel500,
-
-  colors.blueberry400,
-  colors.grapeSoda400,
-  colors.pomegranate400,
-  colors.cinnamon400,
-  colors.cantaloupe400,
-  colors.sourLemon400,
-  colors.greenApple400,
-  colors.jewel400,
-
-  colors.blueberry300,
-  colors.grapeSoda300,
-  colors.pomegranate300,
-  colors.cinnamon300,
-  colors.cantaloupe300,
-  colors.sourLemon300,
-  colors.greenApple300,
-  colors.jewel300,
-
-  colors.blueberry200,
-  colors.grapeSoda200,
-  colors.pomegranate200,
-  colors.cinnamon200,
-  colors.cantaloupe200,
-  colors.sourLemon200,
-  colors.greenApple200,
-  colors.jewel200,
-
-  colors.blueberry100,
-  colors.grapeSoda100,
-  colors.pomegranate100,
-  colors.cinnamon100,
-  colors.cantaloupe100,
-  colors.sourLemon100,
-  colors.greenApple100,
-  colors.jewel100,
-
-  colors.blackPepper600,
-  colors.blackPepper300,
-  colors.blackPepper100,
-  colors.frenchVanilla500,
-  colors.frenchVanilla400,
-  colors.frenchVanilla300,
-  colors.frenchVanilla200,
-  colors.frenchVanilla100,
-];
